Ajout add/edit/remove Sliders

+++ app/Controller/SlidersController.php
@@ -10,21 +10,28 @@ class SlidersController extends AppController {
 		if ($this->request->is('post')) {
 			$d = $this->request->data;
 			$d['Slider']['slider_id'] = null;
-			if ($this->Slider->save($d, true, array('slider_id', 'slider_name', 'slider_desc', 'slider_img'))) {
-				$this->Session->setFlash("L'article à bien été ajouté !", 'notif');
-				$this->redirect(array('controller' => 'sliders', 'action' => 'index', 'admin' => true));
-			} else {
-				$this->Session->setFlash("Un problème est survenu, réessayer !",'notif', array('type' => 'error'));
-				$this->redirect($this->referer());
+			$d['Slider']['slider_img'] = $this->upload_file($d['Slider']['slider_img'], 'sliders');
+			if ($d['Slider']['slider_img'] != false) {
+				if ($this->Slider->save($d, true, array('slider_id', 'slider_name', 'slider_desc', 'slider_img'))) {
+					$this->Session->setFlash("L'article à bien été ajouté !", 'notif');
+					$this->redirect(array('controller' => 'sliders', 'action' => 'index', 'admin' => true));
+				} else {
+					$this->Session->setFlash("Un problème est survenu, réessayer !",'notif', array('type' => 'error'));
+					$this->redirect($this->referer());
+				}
 			}
 		}
 	}
 
-	public function admim_edit($id) {
-		$data = $this->slider->find('first', array('conditions' => array('slider_id' => $id)));
+	public function admin_edit($id) {
+		$data = $this->Slider->find('first', array('conditions' => array('slider_id' => $id)));
 		if ($this->request->is('post') || $this->request->is('put')) {
 			$d = $this->request->data;
+			$this->Slider->id = $id;
+			$d['Slider']['slider_img'] = $d['Slider']['slider_img']['size'] == 0 ? $data['Slider']['slider_img'] : $this->upload_file($d['Slider']['slider_img'], 'sliders');
 			if ($this->Slider->save($d, true, array('slider_name', 'slider_desc', 'slider_img'))) {
+				if ($d['Slider']['slider_img'] != $data['Slider']['slider_img'])
+					$this->delete_file($data['Slider']['slider_img'], 'sliders');
 				$this->Session->setFlash("L'article à bien été supprimé !", 'notif');
 				$this->redirect(array('controller' => 'sliders', 'action' => 'index', 'admin' => true));
 			} else {
@@ -37,9 +44,12 @@ class SlidersController extends AppController {
 
 	public function admin_remove($id) {
 		$this->autoRender = false;
-		$this->Slider->delete($id);
-		$this->Session->setFlash("L'article à bien été supprimé !", 'notif');
-		$this->redirect($this->referer());
+		$data = $this->Slider->find('first', array('conditions' => array('slider_id' => $id)));
+		if ($this->delete_file($data['Slider']['slider_img'], 'sliders')) {
+			$this->Slider->delete($id);
+			$this->Session->setFlash("L'article à bien été supprimé !", 'notif');
+			$this->redirect($this->referer());
+		}
 	}
 
 	public function admin_activated($id) {
